+++ mex/extractors/ff_projects/transform.py
@@ -1,16 +1,18 @@
 from mex.common.exceptions import MExError
-from mex.common.models import ActivityMapping, ExtractedActivity
+from mex.common.models import ActivityMapping, ExtractedActivity, ExtractedOrganization
 from mex.common.types import (
     MergedOrganizationalUnitIdentifier,
     MergedOrganizationIdentifier,
     MergedPersonIdentifier,
     TemporalEntityPrecision,
+    Text,
     YearMonthDay,
 )
 from mex.extractors.ff_projects.models.source import FFProjectsSource
 from mex.extractors.primary_source.helpers import (
     get_extracted_primary_source_id_by_name,
 )
+from mex.extractors.sinks import load
 
 
 def transform_ff_projects_source_to_extracted_activity(
@@ -61,10 +63,11 @@ def transform_ff_projects_source_to_extracted_activity(
             ff_projects_activity.funderOrCommissioner[0].mappingRules[1].forValues or ()
         ):
             continue
-        if org in organization_stable_target_id_by_synonyms:
-            funder_or_commissioner.append(
-                organization_stable_target_id_by_synonyms[org]
-            )
+        funder_or_commissioner = get_or_create_organization(
+            orgs,
+            organization_stable_target_id_by_synonyms,
+        )
+
     if ff_projects_source.rki_az in (
         ff_projects_activity.activityType[0].mappingRules[0].forValues or ()
     ):
@@ -106,3 +109,34 @@ def transform_ff_projects_source_to_extracted_activity(
         hadPrimarySource=get_extracted_primary_source_id_by_name("ff-projects"),
         fundingProgram=ff_projects_source.foerderprogr,
     )
+
+
+def get_or_create_organization(
+    orgs: list[str],
+    extracted_organizations: dict[str, MergedOrganizationIdentifier],
+) -> list[MergedOrganizationIdentifier]:
+    """Get organizations merged ids.
+
+    Args:
+        orgs: organizations from source
+        extracted_organizations: merged organization identifier extracted from wikidata
+
+    Returns:
+        list of matched or created organization identifier
+    """
+    final_organizations: list[MergedOrganizationIdentifier] = []
+    for org in orgs:
+        if wpo := extracted_organizations.get(org):
+            final_organizations.append(wpo)
+        elif org not in ("None", "Not applicable"):
+            new_extracted_organization = ExtractedOrganization(
+                officialName=[Text(value=org)],
+                identifierInPrimarySource=org,
+                hadPrimarySource=get_extracted_primary_source_id_by_name("ff-projects"),
+            )
+            load([new_extracted_organization])
+            final_organizations.append(
+                MergedOrganizationIdentifier(new_extracted_organization.stableTargetId)
+            )
+
+    return final_organizations